refactor: 패키지 이동 & 불필요한 파일 제거
