+++ app/src/main/java/woowacourse/shopping/ui/cart/cartitem/CartItemFragment.kt
@@ -12,6 +12,7 @@ import woowacourse.shopping.data.repository.OrderRepositoryImpl
 import woowacourse.shopping.data.repository.ProductRepositoryImpl
 import woowacourse.shopping.data.repository.RecentProductRepositoryImpl
 import woowacourse.shopping.databinding.FragmentCartItemsBinding
+import woowacourse.shopping.ui.cart.cartitem.adapter.CartAdapter
 import woowacourse.shopping.ui.cart.viewmodel.CartViewModel
 import woowacourse.shopping.ui.cart.viewmodel.CartViewModelFactory
