+++ app/src/main/java/woowacourse/shopping/ui/coupon/uimodel/CouponClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.coupon
+package woowacourse.shopping.ui.coupon.uimodel
 
 interface CouponClickListener {
     fun onCheckBoxClicked(couponId: Long)
