+++ app/src/main/java/woowacourse/shopping/ui/coupon/uimodel/CouponUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.coupon
+package woowacourse.shopping.ui.coupon.uimodel
 
 import java.time.LocalDate
 import java.time.LocalTime
