+++ app/src/main/java/woowacourse/shopping/ui/coupon/uimodel/PaymentInfoUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.coupon
+package woowacourse.shopping.ui.coupon.uimodel
 
 data class PaymentInfoUiModel(
     val orderPrice: Int,
