--- app/src/main/java/woowacourse/shopping/ui/detail/CartButtonClickListener.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.ui.detail
-
-interface CartButtonClickListener {
-    fun onClickAddCartButton()
-}
