+++ app/src/main/java/woowacourse/shopping/ui/detail/uimodel/MostRecentProductClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.detail
+package woowacourse.shopping.ui.detail.uimodel
 
 interface MostRecentProductClickListener {
     fun onClickMostRecentProduct()
