+++ app/src/main/java/woowacourse/shopping/ui/products/ProductContentsActivity.kt
@@ -14,6 +14,8 @@ import woowacourse.shopping.ui.cart.CartActivity
 import woowacourse.shopping.ui.detail.ProductDetailActivity
 import woowacourse.shopping.ui.products.adapter.ProductAdapter
 import woowacourse.shopping.ui.products.adapter.RecentProductAdapter
+import woowacourse.shopping.ui.products.uimodel.LoadingUiModel
+import woowacourse.shopping.ui.products.uimodel.ProductUiModel
 import woowacourse.shopping.ui.products.viewmodel.ProductContentsViewModel
 import woowacourse.shopping.ui.products.viewmodel.ProductContentsViewModelFactory
