+++ app/src/main/java/woowacourse/shopping/ui/products/ProductsMapper.kt
@@ -1,6 +1,7 @@
 package woowacourse.shopping.ui.products
 
 import woowacourse.shopping.domain.model.ProductWithQuantity
+import woowacourse.shopping.ui.products.uimodel.ProductWithQuantityUiModel
 
 fun ProductWithQuantity.toUiModel() =
     ProductWithQuantityUiModel(
