+++ app/src/main/java/woowacourse/shopping/ui/products/uimodel/ProductWithQuantityUiState.kt
@@ -1,7 +1,5 @@
 package woowacourse.shopping.ui.products.uimodel
 
-import woowacourse.shopping.ui.products.ProductUiModel
-
 data class ProductWithQuantityUiState(
     val productWithQuantities: List<ProductUiModel>,
     val isLoading: Boolean = true,
