+++ app/src/main/res/layout/activity_coupon.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="woowacourse.shopping.ui.coupon.CouponViewModel" />
+            type="woowacourse.shopping.ui.coupon.viewmodel.CouponViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
