+++ app/src/main/res/layout/activity_product_detail.xml
@@ -12,7 +12,7 @@
 
         <variable
             name="mostRecentProductClickListener"
-            type="woowacourse.shopping.ui.detail.MostRecentProductClickListener" />
+            type="woowacourse.shopping.ui.detail.uimodel.MostRecentProductClickListener" />
 
     </data>
