+++ app/src/main/res/layout/item_cart.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="cartUiModel"
-            type="woowacourse.shopping.ui.cart.cartitem.CartUiModel" />
+            type="woowacourse.shopping.ui.cart.cartitem.uimodel.CartUiModel" />
 
         <variable
             name="vm"
