+++ app/src/main/res/layout/item_product.xml
@@ -12,7 +12,7 @@
 
         <variable
             name="productWithQuantity"
-            type="woowacourse.shopping.ui.products.ProductWithQuantityUiModel" />
+            type="woowacourse.shopping.ui.products.uimodel.ProductWithQuantityUiModel" />
 
         <variable
             name="addCartClickListener"
